Enforce a maximum file size on the uploads
This is to prevent accidental large file uploads. It applies to both
update and edit.

The edit command can get into a bit of a pickle if it tries to edit
a file that is already to large. In this case it prompts the user
that it may not be able to preform the update.

Fixes #23