Force the API client to render "none" relationships
The API server supports unsetting a relationship using:
PATCH ...
{
  ...,
  "relationships": {
    "assetGroup": { "data" : None }
  }
}
// Or something to this effect

However SimpleJSONAPIClient does not support setting relationships
like that. The JSONAPI specification is kinda vague on this point,
but it looks like a logical operation to preform

It looks like SimpleJSONAPIClient is skipping `nil` as it is falsey.
The quick and dirty fix is to make a version of `nil` which knows
it's JSONAPI relationship is itself (aka None) BUT is also truthy

tl;dr We have a "truthy" version of nil because SimpleJSONAPIClient
      can't false.